// Knowledge/TradeX-MCP/lotte-bridge-main/src/models/request/IGetLoanConfirmRequest.ts
import { Models } from 'tradex-common';
import { ConfirmStatus } from '../../constants/enum';

export interface IGetLoanConfirmRequest extends Models.IDataRequest {
  status: ConfirmStatus;
  fromDate: string;
  toDate: string;
  accountNumber: string;
  subNumber: string;
  nextKey: string;
  stockCode: string;
  fetchCount: number;
}